backend/backend_requirements.txt
absl-py==2.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.114.2
fastapi-cli==0.0.5
filelock==3.16.0
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
jedi==0.19.1
Jinja2==3.1.4
libclang==18.1.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
parso==0.8.4
platformdirs==4.3.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
termcolor==2.4.0
torch==2.4.1
torchtyping==0.1.5
tornado==6.4.1
traitlets==5.14.3
typeguard==2.13.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
x-transformers==1.35.1
